pyproject.toml
[project]
name = "fax-frizzle-lol"
version = "0.1.0"
description = ""
authors = [
    {name = "frizzle-chan"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "cachetools (>=5.5.2,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "python-escpos (>=3.1,<4.0)",
    "requests (>=2.32.3,<3.0.0)",
    "arrow (>=1.3.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "fax_frizzle"}]
package-mode = false


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
autopep8 = "^2.3.2"
pre-commit = "^4.2.0"
isort = "^6.0.1"
flake8-bugbear = "^24.12.12"
mypy = "^1.15.0"
types-requests = "^2.32.0.20250328"
types-cachetools = "^5.5.0.20240820"
pytest = "^8.3.5"
pytest-regressions = {extras = ["image"], version = "^2.7.0"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = ["escpos.*"]
follow_untyped_imports = true